+++ packages/yuudachi/src/functions/anti-raid/filters.ts
@@ -4,8 +4,8 @@ import type RE2 from 're2';
 
 export function joinFilter(
 	member: GuildMember,
-	joinFrom: number = Number.NEGATIVE_INFINITY,
-	joinTo: number = Number.POSITIVE_INFINITY,
+	joinFrom = Number.NEGATIVE_INFINITY,
+	joinTo = Number.POSITIVE_INFINITY,
 ): boolean {
 	if (!member.joinedTimestamp) return false;
 	return member.joinedTimestamp <= joinFrom || member.joinedTimestamp >= joinTo;
@@ -13,8 +13,8 @@ export function joinFilter(
 
 export function ageFilter(
 	member: GuildMember,
-	ageFrom: number = Number.NEGATIVE_INFINITY,
-	ageTo: number = Number.POSITIVE_INFINITY,
+	ageFrom = Number.NEGATIVE_INFINITY,
+	ageTo = Number.POSITIVE_INFINITY,
 ): boolean {
 	return member.user.createdTimestamp <= ageFrom || member.user.createdTimestamp >= ageTo;
 }
